fix(bar): fix on bar tooglip grouped false
- Add condition to execute $$.unexpandCircles()
- Split test's util file to be helper/util for ESM test

Ref #https://github.com/naver/billboard.js/issues/1663#issuecomment-690991679
